Some more cleanup
