+++ polyglot/piranha/src/models/piranha_arguments.rs
@@ -31,14 +31,6 @@ use super::{
   language::PiranhaLanguage,
 };
 
-pub enum PiranhaInput {
-  CommandLineInput,
-  API {
-    path_to_codebase: String,
-    path_to_configurations: String,
-    dry_run: bool,
-  },
-}
 
 /// A refactoring tool that eliminates dead code related to stale feature flags.
 #[derive(Deserialize, Clone, Builder, Getters, CopyGetters, Debug, Parser, Default)]
@@ -151,12 +143,12 @@ impl PiranhaArguments {
     self.language[0].clone()
   }
 
-  fn get_path_to_piranha_arguments_toml(&self) -> PathBuf {
+  pub(crate) fn get_path_to_piranha_arguments_toml(&self) -> PathBuf {
     PathBuf::from(self.path_to_configurations.to_string())
       .join(default_name_of_piranha_argument_toml())
   }
 
-  fn new(path_to_piranha_arguments_toml: PathBuf) -> Self {
+  pub(crate) fn new(path_to_piranha_arguments_toml: PathBuf) -> Self {
     let args: PiranhaArguments = read_toml(&path_to_piranha_arguments_toml, false);
     let input_substitutions = args.substitutions();
     let piranha_language = PiranhaLanguage::from(args.get_language().as_str());
@@ -176,7 +168,7 @@ impl PiranhaArguments {
     }
   }
 
-  fn merge(&self, other: PiranhaArguments) -> Self {
+  pub(crate) fn merge(&self, other: PiranhaArguments) -> Self {
     Self {
       path_to_code_base: Self::_merge(
         self.path_to_code_base.clone(),
@@ -204,6 +196,7 @@ impl PiranhaArguments {
         default_path_to_output_summaries(),
       ),
       language: Self::_merge(self.language.clone(), other.language, default_languages()),
+
       piranha_language: Self::_merge(
         self.piranha_language.clone(),
         other.piranha_language,
@@ -244,23 +237,8 @@ impl PiranhaArguments {
   }
 }
 
-impl From<PiranhaInput> for PiranhaArguments {
-  fn from(input: PiranhaInput) -> Self {
-    let input_opts = match input {
-      PiranhaInput::CommandLineInput => PiranhaArguments::parse(),
-      PiranhaInput::API {
-        path_to_codebase,
-        path_to_configurations,
-        dry_run,
-      } => PiranhaArgumentsBuilder::default()
-        .path_to_code_base(path_to_codebase)
-        .path_to_configurations(path_to_configurations)
-        .path_to_output_summaries(None)
-        .dry_run(dry_run)
-        .build()
-        .unwrap(),
-    };
-    let piranha_argument = PiranhaArguments::new(input_opts.get_path_to_piranha_arguments_toml());
-    return input_opts.merge(piranha_argument);
-  }
+
+pub trait Merge {
+  /// Merge another object into this object.
+  fn merge(&mut self, other: Self, default: Self);
 }
